Routing-pattern/requirements.txt
ollama==0.1.7
python-dotenv==1.0.0
pydantic==2.5.0
pytest==7.4.3
pytest-asyncio==0.21.1
streamlit==1.20.0
